// src/db/database.types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export interface Database {
  graphql_public: {
    Tables: Record<never, never>;
    Views: Record<never, never>;
    Functions: {
      graphql: {
        Args: {
          extensions?: Json;
          operationName?: string;
          query?: string;
          variables?: Json;
        };
        Returns: Json;
      };
    };
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
  public: {
    Tables: {
      instrument_value_changes: {
        Row: {
          after_value_grosze: number;
          before_value_grosze: number;
          created_at: string;
          id: string;
          instrument_id: string;
        };
        Insert: {
          after_value_grosze: number;
          before_value_grosze: number;
          created_at?: string;
          id?: string;
          instrument_id: string;
        };
        Update: {
          after_value_grosze?: number;
          before_value_grosze?: number;
          created_at?: string;
          id?: string;
          instrument_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "instrument_value_changes_instrument_id_fkey";
            columns: ["instrument_id"];
            isOneToOne: false;
            referencedRelation: "instruments";
            referencedColumns: ["id"];
          },
        ];
      };
      instruments: {
        Row: {
          created_at: string;
          current_value_grosze: number;
          deleted_at: string | null;
          goal_grosze: number | null;
          id: string;
          invested_money_grosze: number;
          name: string;
          owner_id: string;
          short_description: string | null;
          type: Database["public"]["Enums"]["instrument_type"];
          updated_at: string;
          wallet_id: string;
        };
        Insert: {
          created_at?: string;
          current_value_grosze: number;
          deleted_at?: string | null;
          goal_grosze?: number | null;
          id?: string;
          invested_money_grosze: number;
          name: string;
          owner_id: string;
          short_description?: string | null;
          type: Database["public"]["Enums"]["instrument_type"];
          updated_at?: string;
          wallet_id: string;
        };
        Update: {
          created_at?: string;
          current_value_grosze?: number;
          deleted_at?: string | null;
          goal_grosze?: number | null;
          id?: string;
          invested_money_grosze?: number;
          name?: string;
          owner_id?: string;
          short_description?: string | null;
          type?: Database["public"]["Enums"]["instrument_type"];
          updated_at?: string;
          wallet_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "instruments_wallet_id_fkey";
            columns: ["wallet_id"];
            isOneToOne: false;
            referencedRelation: "wallets";
            referencedColumns: ["id"];
          },
        ];
      };
      wallets: {
        Row: {
          created_at: string;
          deleted_at: string | null;
          description: string | null;
          id: string;
          name: string;
          owner_id: string;
          updated_at: string;
        };
        Insert: {
          created_at?: string;
          deleted_at?: string | null;
          description?: string | null;
          id?: string;
          name: string;
          owner_id: string;
          updated_at?: string;
        };
        Update: {
          created_at?: string;
          deleted_at?: string | null;
          description?: string | null;
          id?: string;
          name?: string;
          owner_id?: string;
          updated_at?: string;
        };
        Relationships: [];
      };
    };
    Views: Record<never, never>;
    Functions: Record<never, never>;
    Enums: {
      instrument_type: "bonds" | "etf" | "stocks";
    };
    CompositeTypes: Record<never, never>;
  };
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"] | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never;